+++ R/tm_data_table.R
@@ -325,7 +325,8 @@ ui_data_table <- function(id,
 }
 
 
-#' @importFrom dplyr count_
+#' @importFrom dplyr count
+#' @importFrom rlang !!! syms
 #' @importFrom DT datatable
 srv_data_table <- function(input,
                            output,
@@ -353,7 +354,7 @@ srv_data_table <- function(input,
     validate(need(all(variables %in% names(df)), "not all selected variables exist"))
 
     dataframe_selected <- if (if_distinct()) {
-      count_(df, variables)
+      count(df, !!!syms(variables))
     } else {
       df[variables]
     }