+++ NEWS.md
@@ -15,12 +15,11 @@
 ### Miscellaneous
 * Sparklines no longer shown for numeric variables with more than 100000 rows.
 * Colors legend for each plot.
-* Remove "Show variables other than in ADSL" checkbox in variable browser module.
+* Removed "Show variables other than in ADSL" checkbox in variable browser module.
 * Capitalize default module labels.
 * Some of the outputs of `tm_missing_data` are conditional to the data being a CDISC data.
-* Add `pre_output` and `post_output` arguments to `tm_data_table`, `tm_missing_data`, `tm_variable_browser`.
+* Added `pre_output` and `post_output` arguments to `tm_data_table`, `tm_missing_data`, `tm_variable_browser`.
 * Fixed issue in `tm_data_table` when selecting "show distinct rows" if variable names had non-alphanumeric characters. 
-
 
 # teal.modules.general 0.2.7