+++ NEWS.md
@@ -12,6 +12,7 @@
 
 ### Bug Fixes
 * Fixed edge case errors in `tm_missing_data` caused by input `dataset` not having categorical variables or being a `data.frame` instead of `tibble`.
+* Fixed `tm_a_pca` biplot plot when coloring with a factor/character/numeric with < 6 values variable.
 
 ### Miscellaneous
 * Sparklines no longer shown for numeric variables with more than 100000 rows.