+++ NEWS.md
@@ -1,3 +1,6 @@
+# teal.modules.general 0.2.8.9000
+Issues with no news:
+
 # teal.modules.general 0.2.8
 
 ### New Module